+++ hospitapp/src/repositories/ips_mongo.repository.interfaces.ts
